Updated js, constructor, lights editor
Basically I made javascript responsible for the rendering of the cars and traffic lights, so python just sends data.  I guess we don't need an attribute navigator but we'll see. If you press on a car you'll get it's info, the same applies for the traffic lights (during pause).  Made a couple of other changes. Also fixed the problem with lights in the corner.....( Нема більше гірлянди(